requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
requests
faster-whisper
gtts
langgraph
pytesseract
pillow
python-multipart
python-telegram-bot
